Narrow CharsXY to u16 (from usize)
Using usize to represent character screen coordinates was overkill, and
especially now that the graphical coordinates are capped at i16.  To
make things more homogeneous, cap CharsXY to u16.

Note that PixelsXY allows negative numbers to support partial off-screen
rendering, which is very useful, but CharsXY does not.  The reason is
that the console primitives do not allow off-screen printing and thus
we shouldn't be worrying about negative numbers everywhere.

+++ std/src/console/readline.rs
@@ -45,7 +45,7 @@ async fn read_line_interactive(
         // Assumes that the prompt was printed at column 0.  If that was not the case, line length
         // calculation does not work.
         let console_size = console.size()?;
-        console_size.x - prompt.len()
+        usize::from(console_size.x) - prompt.len()
     };
 
     // Insertion position *within* the line, without accounting for the prompt.
